% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidytuesday_dictionary.R
\name{create_tidytuesday_dictionary}
\alias{create_tidytuesday_dictionary}
\title{Create the TidyTuesday data dictionary}
\usage{
create_tidytuesday_dictionary(x)
}
\description{
Create the TidyTuesday data dictionary
}
