% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidytuesday_folder.R
\name{create_tidytuesday_folder}
\alias{create_tidytuesday_folder}
\title{Create the TidyTuesday directory}
\usage{
create_tidytuesday_folder()
}
\description{
Create the TidyTuesday directory
}
