% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidytuesday_dictionary.R
\name{create_tt_dict}
\alias{create_tt_dict}
\title{Create the TidyTuesday data dictionary}
\usage{
create_tt_dict(x)
}
\arguments{
\item{x}{an in memory data.frame or tibble.}
}
\description{
Create the TidyTuesday data dictionary
}
