% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tidytuesday_readme.R
\name{create_tt_readme}
\alias{create_tt_readme}
\title{Color scale constructor for tomtom colors}
\usage{
create_tt_readme(title, ...)
}
\arguments{
\item{title}{datasets for creating the data dictionary and flat files}
}
\description{
Color scale constructor for tomtom colors
}
