% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_table.R
\name{fill_table}
\alias{fill_table}
\title{Fill table for readme}
\usage{
fill_table(title, swap = FALSE)
}
\arguments{
\item{title}{title of week's data}

\item{swap}{append data/year or just date}
}
\value{
kable table
}
\description{
Fill table for readme
}
