% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_type.R
\name{get_data_type}
\alias{get_data_type}
\title{Update the data_type.csv file for tidytuesdayR}
\usage{
get_data_type(file_names, delimiter)
}
\description{
Update the data_type.csv file for tidytuesdayR
}
