% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_tidytuesday.R
\name{post_tidytuesday}
\alias{post_tidytuesday}
\title{Create the R4DS Community TidyTuesday posts}
\usage{
post_tidytuesday(exploring, short_link)
}
\description{
Create the R4DS Community TidyTuesday posts
}
