% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_mastodon.R
\name{post_tt_toot}
\alias{post_tt_toot}
\title{Create the R4DS Community TidyTuesday posts}
\usage{
post_tt_toot(exploring, short_link, alt_add)
}
\description{
Create the R4DS Community TidyTuesday posts
}
