% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_type.R
\name{update_data_type}
\alias{update_data_type}
\title{Update the data_type.csv file for tidytuesdayR}
\usage{
update_data_type(file_names, delimiter)
}
\description{
Update the data_type.csv file for tidytuesdayR
}
