% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_mastodon.R
\name{post_tt_toot}
\alias{post_tt_toot}
\title{Create the R4DS Community TidyTuesday posts}
\usage{
post_tt_toot(exploring, short_link, alt_add)
}
\arguments{
\item{exploring}{string indicating the topic}

\item{short_link}{string indicating the link to the article used for this week.}

\item{alt_add}{A length two vector of strings adding alt text for the included images.}
}
\description{
Create the R4DS Community TidyTuesday posts
}
