% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tidytuesday_template.R
\name{use_tidytuesday_readme}
\alias{use_tidytuesday_readme}
\title{Use a usethis-style template for TidyTuesdy
Thank you Jenny + Sharla + Sharleen!}
\usage{
use_tidytuesday_readme(title_subject, dataset_name)
}
\arguments{
\item{template}{Path to template file relative to `templates/`}

\item{save_as}{Path of file to create, relative to root of active project.
Defaults to `template`}

\item{data}{A list of data passed to the template.}

\item{open}{Open the newly created file for editing?}
}
\value{
A logical vector indicating if file was modified.
}
\description{
Creates a file from data and a template.
Combine your template file with your data.
}
\examples{

}
